[PVR] Refactor CPVRGUIActions step 6: Move power management related actions to own class.
